+++ lib/src/dtos/collection_field.dart
@@ -5,8 +5,6 @@ import "package:json_annotation/json_annotation.dart";
 import "../caster.dart" as caster;
 import "jsonable.dart";
 
-part "collection_field.g.dart"; // not actually used
-
 /// Response DTO of a single collection schema field.
 @JsonSerializable(explicitToJson: true)
 class CollectionField implements Jsonable {